refactor(otto/pond): rename to river
